pyproject.toml
[tool.poetry]
name = "gaudrail"
version = "0.1.0"
description = ""
authors = ["Toufik Al Khawli"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11.9,<3.12.dev0"
langchain-openai = "^0.3.6"
guardrails-ai = "^0.6.3"
numpy = "^2.2.3"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
streamlit = "^1.42.2"
nest-asyncio = "^1.6.0"
transformers = "^4.36.0"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^8.3.4"
pytest-mock = "^3.14.0"
tox = "^3.28.0"
tox-poetry = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
